+++ HISTORY.md
@@ -2,6 +2,8 @@
 
 ## Version 0.0.9 02 September 2026
 1. Load last partner's greeting on startup
+2. Items being added to history when selected
+
 
 ## Version 0.0.8 01 September 2026
 1. Implement message.py as dynamic data repository